requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# OpenAI integration
openai>=1.0.0

# Database (Postgres)
psycopg2-binary>=2.9.0

# Web scraping (for Walmart integration)
beautifulsoup4>=4.12.0
google-search-results>=2.4.2

# Testing (optional)
pytest>=7.4.0




